Client'tan veri ekleme operasyonu ve bazı tasarımsal iyileştirmeler yapıldı.

+++ ETradeAPI/Infrastructure/ETradeAPI.Persistence/Migrations/20230505141332_mig_1.cs
@@ -32,7 +32,7 @@ namespace ETradeAPI.Persistence.Migrations
                     Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Description = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Stock = table.Column<long>(type: "bigint", nullable: false),
+                    Stock = table.Column<int>(type: "int", nullable: false),
                     Price = table.Column<float>(type: "real", nullable: false),
                     CreatedDate = table.Column<DateTime>(type: "datetime2", nullable: false),
                     UpdatedDate = table.Column<DateTime>(type: "datetime2", nullable: false)
