+++ ETradeAPI/Infrastructure/ETradeAPI.Persistence/Migrations/ETradeDbContextModelSnapshot.cs
@@ -93,8 +93,8 @@ namespace ETradeAPI.Persistence.Migrations
                     b.Property<float>("Price")
                         .HasColumnType("real");
 
-                    b.Property<long>("Stock")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Stock")
+                        .HasColumnType("int");
 
                     b.Property<DateTime>("UpdatedDate")
                         .HasColumnType("datetime2");
